+++ packages/css/src/scales/css-common.ts
@@ -1,4 +1,8 @@
-export type { Globals } from 'csstype'
+import type { Globals as CSSGlobals } from 'csstype'
+
+// We add null and false values for convenience.
+// `color: isActive && 'blue'` syntax common in JS projects
+export type Globals = CSSGlobals | null | false
 
 // DataType is not exported from `csstype`, and we need to do some slight
 // modifications to it for performance and correctness