+++ packages/css/test/index.ts
@@ -54,8 +54,8 @@ const theme: Theme = {
   },
   opacities: [0, '50%'],
   transitions: {
-    standard: '0.3s ease-in-out'
-  }
+    standard: '0.3s ease-in-out',
+  },
 }
 
 test('returns a function', () => {
@@ -118,7 +118,7 @@ test('returns nested responsive styles', () => {
     color: 'primary',
     h1: {
       py: [3, 4],
-      scrollPaddingY: [2, 4]
+      scrollPaddingY: [2, 4],
     },
   })({ theme })
   expect(result).toEqual({
@@ -133,7 +133,6 @@ test('returns nested responsive styles', () => {
         paddingBottom: 32,
         scrollPaddingBottom: 32,
         scrollPaddingTop: 32,
-  
       },
     },
   })
@@ -516,12 +515,28 @@ test('returns correct media query order 2', () => {
     'paddingRight',
     'paddingTop',
     'paddingBottom',
-    'scrollPadding'
+    'scrollPadding',
   ])
 })
 
 test('supports vendor properties', () => {
   expect(css({ WebkitOverflowScrolling: 'touch' })(theme)).toStrictEqual({
     WebkitOverflowScrolling: 'touch',
+  })
+})
+
+test('filters `false` and `null`', () => {
+  const isActive = false
+  const isAwesome = true
+
+  expect(
+    css({
+      backgroundColor: 'background',
+      color: isActive && 'blue',
+      borderColor: isAwesome && 'blue',
+    })(theme)
+  ).toStrictEqual({
+    backgroundColor: 'white',
+    borderColor: 'blue',
   })
 })